factor out solc JSON interface crate
