+++ Cargo.toml
@@ -24,6 +24,7 @@ lld-sys = { version = "0.1.0-dev.12", path = "crates/lld-sys" }
 revive-llvm-context = { version = "0.1.0-dev.12", path = "crates/llvm-context" }
 revive-runtime-api = { version = "0.1.0-dev.12", path = "crates/runtime-api" }
 revive-runner = { version = "0.1.0-dev.12", path = "crates/runner" }
+revive-solc-json-interface = { version = "0.1.0-dev.12", path = "crates/solc-json-interface" }
 revive-solidity = { version = "0.1.0-dev.12", path = "crates/solidity" }
 revive-stdlib = { version = "0.1.0-dev.12", path = "crates/stdlib" }
 revive-build-utils = { version = "0.1.0-dev.12", path = "crates/build-utils" }
