+++ crates/solc-json-interface/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "revive-solc-json-interface"
+version.workspace = true
+authors.workspace = true
+license.workspace = true
+edition.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+description = "Rust bindings for the solc standard JSON and combined JSON interface"
+
+[features]
+default = ["parallel"]
+parallel = ["rayon"]
+resolc = []
+
+[dependencies]
+revive-common = { workspace = true }
+
+anyhow = { workspace = true }
+rayon = { workspace = true, optional = true }
+semver = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+serde_stacker = { workspace = true }
