+++ crates/solc-json-interface/src/standard_json/input/settings/metadata_hash.rs
