+++ crates/solc-json-interface/src/standard_json/input/settings/optimizer/details.rs
